rename special directory `for-pages` to `_pages`
