+++ packages/web/src/components/_pages/liquidity/Statistics/index.spec.tsx
