+++ packages/web/src/components/_pages/liquidity/TokenSymbol/index.tsx
