+++ packages/web/src/fixtures/_pages/liquidity/constants/number.ts
