+++ packages/web/src/fixtures/_pages/liquidity/uniswap-pool/__mocks__/hooks.ts
