+++ packages/web/src/fixtures/_pages/liquidity/uniswap-pool/abi.ts
