+++ packages/web/src/fixtures/_pages/liquidity/uniswap-pool/cache-path.ts
