+++ packages/web/src/fixtures/_pages/liquidity/uniswap-pool/client.ts
